refactor: Extract getSlugFromInstitutionLabel
