+++ packages/cozy-doctypes/src/banking/matching-accounts.js
@@ -1,16 +1,6 @@
 const sortBy = require('lodash/sortBy')
 const { eitherIncludes } = require('./matching-tools')
-const labelSlugs = require('./label-slugs')
-
-const institutionLabelsCompiled = Object.entries(labelSlugs).map(
-  ([ilabelRx, slug]) => {
-    if (ilabelRx[0] === '/' && ilabelRx[ilabelRx.length - 1] === '/') {
-      return [new RegExp(ilabelRx.substr(1, ilabelRx.length - 2), 'i'), slug]
-    } else {
-      return [ilabelRx, slug]
-    }
-  }
-)
+const { getSlugFromInstitutionLabel } = require('./slug-account')
 
 const findExactMatch = (attr, account, existingAccounts) => {
   const sameAttr = existingAccounts.filter(
@@ -92,19 +82,6 @@ const creditCardMatch = (account, existingAccount) => {
   const other = ccAccount === account ? existingAccount : account
   if (other.number.slice(-4) === lastDigits) {
     return true
-  }
-}
-
-const getSlugFromInstitutionLabel = institutionLabel => {
-  for (let [rx, slug] of institutionLabelsCompiled) {
-    if (rx instanceof RegExp) {
-      const match = institutionLabel.match(rx)
-      if (match) {
-        return slug
-      }
-    } else if (rx.toLowerCase() === institutionLabel.toLowerCase()) {
-      return slug
-    }
   }
 }
